Update round complete popup. Add change difficulty, show points earnt

+++ src/main/java/seng201/team53/items/towers/Tower.java
@@ -78,7 +78,7 @@ public class Tower implements Item, Upgradeable {
 
         GameDifficulty difficulty = GameEnvironment.getGameEnvironment().getDifficulty();
         long reloadSpeed = type.getReloadSpeed().toMillis();
-        reloadSpeed /= (long)difficulty.getTowerReloadModifier();
+        reloadSpeed /= difficulty.getTowerReloadModifier();
         reloadSpeed /= reloadSpeedModifier;
 
         long deltaTime = System.currentTimeMillis() - getLastGenerateTime();
